+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Upcoming release
+
+- Session close events now report the time when they happened instead of when they're sent
+
 ## 3.4.0
 
 - Fix device's last registered date encoding